DAT-1056-refactor-responses
