+++ .pylintrc
@@ -338,7 +338,7 @@ ignore-docstrings=yes
 ignore-imports=no
 
 # Minimum lines number of a similarity.
-min-similarity-lines=4
+min-similarity-lines=6
 
 
 [MISCELLANEOUS]
