+++ cmdb/database/utils.py
@@ -30,9 +30,14 @@ from bson.tz_util import utc
 
 from cmdb.framework.rendering.render_result import RenderResult
 from cmdb.cmdb_objects.cmdb_dao import CmdbDAO
+from cmdb.importer.importer_response import BaseImporterResponse, ImportMessage
+from cmdb.media_library.media_file_base import MediaFileManagementBase
+from cmdb.security.auth.auth_settings import AuthSettingsDAO
+from cmdb.security.auth.provider_config import AuthProviderConfig
 from cmdb.settings.date.date_settings import DateSettingsDAO
 from cmdb.search.search_result import SearchResult, SearchResultMap
 from cmdb.docapi.docapi_template.docapi_template_base import TemplateManagementBase
+from cmdb.user_management.models.right import BaseRight
 # -------------------------------------------------------------------------------------------------------------------- #
 
 LOGGER = logging.getLogger(__name__)
@@ -96,6 +101,12 @@ def default(obj):
     if isinstance(obj, (CmdbDAO,
                         RenderResult,
                         TemplateManagementBase,
+                        AuthSettingsDAO,
+                        ImportMessage,
+                        BaseImporterResponse,
+                        MediaFileManagementBase,
+                        AuthProviderConfig,
+                        BaseRight,
                         DateSettingsDAO)):
         return obj.__dict__
