+++ cmdb/exportd/__init__.py
@@ -1,3 +1,4 @@
+
 # DATAGERRY - OpenSource Enterprise CMDB
 # Copyright (C) 2024 becon GmbH
 #
@@ -13,13 +14,3 @@
 #
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
-"""TODO: document"""
-
-from cmdb.exportd.exportd_job.exportd_job import ExportdJob
-from cmdb.exportd.exportd_logs.exportd_log import ExportdMetaLog
-# -------------------------------------------------------------------------------------------------------------------- #
-
-__COLLECTIONS__ = [
-    ExportdJob,
-    ExportdMetaLog
-    ]
