+++ cmdb/exportd/constants.py
@@ -13,3 +13,13 @@
 #
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
+"""TODO: document"""
+
+from cmdb.exportd.exportd_job.exportd_job import ExportdJob
+from cmdb.exportd.exportd_logs.exportd_log import ExportdMetaLog
+# -------------------------------------------------------------------------------------------------------------------- #
+
+__COLLECTIONS__ = [
+    ExportdJob,
+    ExportdMetaLog
+    ]
