+++ cmdb/framework/__init__.py
@@ -13,34 +13,3 @@
 #
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
-"""
-This module represents the core system of the CMDB.
-All initializations and specifications for creating objects,
-object types and their fields are controlled here.
-Except for the managers, this module can be used completely modular.
-The respective DAO is used to apply the attributes and to convert
-the elements for the database.
-"""
-from cmdb.cmdb_objects.cmdb_dao import CmdbDAO
-from cmdb.cmdb_objects.cmdb_object import CmdbObject
-from cmdb.cmdb_objects.cmdb_location import CmdbLocation
-from cmdb.cmdb_objects.cmdb_section_template import CmdbSectionTemplate
-
-from cmdb.framework.models.type import TypeModel
-from cmdb.framework.models.category import CategoryModel
-from cmdb.framework.models.link import ObjectLinkModel
-from cmdb.framework.models.log import CmdbLog, CmdbObjectLog, CmdbMetaLog
-# -------------------------------------------------------------------------------------------------------------------- #
-
-CmdbLog.register_log_type(CmdbObjectLog.__name__, CmdbObjectLog)
-
-# List of init collections
-__COLLECTIONS__ = [
-    CmdbObject,
-    TypeModel,
-    CategoryModel,
-    CmdbMetaLog,
-    ObjectLinkModel,
-    CmdbLocation,
-    CmdbSectionTemplate
-]
