+++ cmdb/interface/rest_api/importer_routes/importer_object_routes.py
@@ -42,12 +42,12 @@ from cmdb.interface.rest_api.importer_routes.importer_route_utils import get_fil
 from cmdb.user_management.models.user import UserModel
 from cmdb.manager.manager_provider import ManagerType, ManagerProvider
 from cmdb.framework.rendering.cmdb_render import CmdbRender
-from cmdb.importer import load_parser_class,\
-                          load_importer_class,\
-                          load_importer_config_class,\
-                          __OBJECT_IMPORTER__,\
-                          __OBJECT_PARSER__, \
-                          __OBJECT_IMPORTER_CONFIG__
+from cmdb.importer.importer_helper import load_parser_class,\
+                                          load_importer_class,\
+                                          load_importer_config_class,\
+                                          __OBJECT_IMPORTER__,\
+                                          __OBJECT_PARSER__, \
+                                          __OBJECT_IMPORTER_CONFIG__
 
 from cmdb.errors.manager import ManagerInsertError
 from cmdb.errors.security import AccessDeniedError
